Account for inheritance with custom type hint
When using a custom type hint on a collection and generate
an `add` method, reflection on the object will not be able
to access the private property of the parent class and when
using reflection on the class type of the relation an exception
will be thrown by PHP7.1 when the added object is of the
type hinted type and does not implement or extends the real
type, although the real type may implement the type hint type.

See `test/Generator/PracticalVehicleOwnerTest.php` for the
different cases and how they are handled.